Ability to check whether an optional value is NULL
There was a need to check whether the Optional is NULL. The library is built in such a way that it is impossible to reach such a check because all this logic is hidden in the internal package. Therefore, a simple proposal to add 2 functions that allow one to do this.